Dispatch.java: Class that execute every command ":.uno" that the libreoffice support
SpeechPropertiesCreator.java: has no modifications

CoGrOO.java: Class that implement the gramatical corrector

KeyEvent.java: Class that implement a keyboard listener and change between the dictation mode and command mode. the switch key is "Alt Graph".

Numbers.java:New Numbers are added and now if none number are recognized the class return "-1"

TableNames.java: The Class that implement the tables of font names

TableGramatical.java: The class that implement the table of suggetions of the gramatical corrector

PrintAndSave.java: The class that implement a function of save and print, but the print method is not complete.